feat(local-runner): run orchestration pipelines client-side, with no server
The orchestration engine needed a Node process only because its runners did: one
spawns a Python kernel, both read notebooks off disk. The engine itself is pure
control flow, so this splits the two apart.

- orchestrate-core.ts: the engine — graph, dependencies, gates, events, output
  helpers — with the runner supplied as an OrchestrationStepExecutor. No node:
  imports are reachable from it.
- orchestrate.ts: the Node binding. Same public API; orchestrate() is unchanged.
- orchestrate-cloud.ts: a fetch-only cloud executor. orchestrateInCloud() runs
  the identical workflow callback in a browser against Deepnote Cloud.
- A self-contained orchestrator.iife.js bundle a static page can <script> in.
- examples/local-runner/client-orchestration: fan-out -> quality gate -> arbiter,
  running entirely in the page.

Two constraints follow from having no server, and both fail with an explanation
rather than deep inside an API call: notebooks must already exist (steps name a
notebookId; a viewer-scoped token may run a notebook, not create one), and there
is no local target (a browser has no Python kernel).
